# perf(stdlib): ⚡️  implement manifestJsonEx as native Go