```
❯ make benchmark FILTER=Builtin_manifestJsonEx
go build ./cmd/jsonnet
./benchmark.sh Builtin_manifestJsonEx
Running Before Test... (10s)
Running After Test... (10s)
benchmark                               old ns/op     new ns/op     delta
Benchmark_Builtin_manifestJsonEx-16     22656394      7502016       -66.89%
```

---

EDIT
~I could use a small amount of help here. The test is currently failing seemingly because the output doesn't match. I suspect it's some sort of whitespace issue, but I can't tell.~

Fixed the problem, I was setting an initial indentation.