@ghostsquad 2020-06-12
This should resolve this issue:
https://github.com/coreos/kube-prometheus/issues/537

on version 1.16.0 - takes 19s to render
after update - takes 2s to render