@sbarzowski 2020-06-13
Releases are not coupled with features, they are cut every couple of months. The last one was on May 22. Release process takes quite a lot of time (take a look at the release procedure: https://github.com/google/jsonnet/pull/802/files). I would like to automate some of that stuff eventually, but it's not easy (most steps may uncover problems requiring manual intervention). 

Anyway, depending on `master` version should be fairly safe, so for bleeding edge I recommend using that.