@ghostsquad 2020-06-15
@sbarzowski holy cow. I had no idea that releases were that involved. Ok, I'll continue to recommend folks use the `go get -u` method of installing. If I'm not mistaken, that uses the latest on `master` not the releases?