@sbarzowski 2020-06-15
> If I'm not mistaken, that uses the latest on master not the releases?

Yes, I think so.